updated readme
When trying to setup a test to see how everything worked, I encountered
a null error.  I copied the example from the Readme.  After reading the
code I added the other two fields for the Server struct and the error
went away.  I added the other two fields to the documenation rather than
changing code to catch the errors, it was faster.